requirements.txt
pandas
seaborn
numpy
matplotlib
scikit-learn
catboost
xgboost
dill
flask
datasets
lightgbm
shap
statsmodels
econml
azure-storage-blob
torch
torchvision
torchaudio
transformers
pytorchvideo
librosa
opencv-python
joblib
python-dotenv